requirements.txt
awscli==1.36.4
boto3==1.35.63
botocore==1.35.63
colorama==0.4.6
docutils==0.16
jmespath==1.0.1
pyasn1==0.6.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
rsa==4.7.2
s3transfer==0.10.3
six==1.16.0
urllib3==2.2.3
